Snapper UI moves to the snapper_ai package; monitor keeps the providers
Companion to snapper-ai c4868ed. The views, series assembly, and
templates leave monitor_app; what stays is exactly the experiment-
specific half, now in monitor_app/snapper_providers.py registered from
AppConfig.ready(): the epicprod and testbed ScopeProviders (curve
extraction, labels, curve families, panda/workflow/datataking card
builders, the testbed run-arc machinery), the host card template
(_snapper_cards.html) carrying the monitor page links, and the host
hooks (UserPreference-backed prefs, SysConfig values, scheduler
status, system page URL). Reference resolution (snapper_resolvers) and
the MCP/REST layers are unchanged.

Snapper mounts at the project level (public paths unchanged); the nav
reverse moves to the snapper_ai namespace. The lightweight deploy
gains a snapper-ai sync stanza on the pcs pattern.

Verified: report/system/cut render on both scopes with structure
identical to pre-move baselines; window arrows and de-noised labels
intact; provider cards render with stats, chips, and references.
