--- src/monitor_app/templates/monitor_app/_snapper_cut.html
@@ -1,71 +1,0 @@
-{% load swf_fmt %}
-{% if error %}
-<div class="text-danger">{{ error }}</div>
-{% else %}
-<style>
-.snapper-cut-card { border:1px solid var(--bs-border-color,#ccc);
-                    border-radius:.4em; padding:.6em .8em; }
-</style>
-<p class="mb-2">
-  State at <strong>{{ requested_at|fmt_dt }}</strong>
-  {% if snap_age_text %}<span class="text-muted">· snap <span title="{{ actual_snap_time|fmt_dt }}">{{ snap_age_text }}</span> earlier</span>{% endif %}
-</p>
-{% if coverage_notice %}
-<p class="mb-2">
-  <span class="snapper-chip" style="background:{{ coverage_notice.chip.color }}">{{ coverage_notice.label }}</span>
-  {{ coverage_notice.detail }}
-</p>
-{% endif %}
-<div class="d-flex flex-wrap gap-3 align-items-stretch">
-  {% for card in cards %}
-  <div class="snapper-cut-card {% if card.changed %}border-2{% endif %}">
-    <div class="d-flex align-items-baseline gap-2 mb-1">
-      <strong>{{ card.name }}</strong>
-      {% if card.kind == 'health' %}
-      <span class="snapper-chip" style="background:{{ card.chip.color }}">{{ card.chip.value }}</span>
-      {% endif %}
-      {% if card.changed %}<span class="badge bg-info text-dark">changed in this snap</span>{% endif %}
-      <span class="text-muted small"{% if card.assessed_age_text %} title="assessed {{ card.assessed_at|fmt_dt }}"{% endif %}>assessed
-        {% if card.assessed_age_text %}{{ card.assessed_age_text }} earlier{% else %}{{ card.assessed_at|fmt_dt }}{% endif %}</span>
-    </div>
-
-    {% if card.kind == 'health' %}
-      {% if card.reason %}<div class="mb-1">{{ card.reason }}</div>{% endif %}
-      {% if card.non_ok_checks %}
-      <table class="table table-sm mb-1 w-auto">
-        {% for check in card.non_ok_checks %}
-        <tr>
-          <td><a href="{% url 'monitor_app:system_status' %}">{{ check.name }}</a></td>
-          <td><span class="snapper-chip" style="background:{{ check.chip.color }}">{{ check.chip.value }}</span></td>
-          <td>{{ check.summary }}</td>
-        </tr>
-        {% endfor %}
-      </table>
-      {% else %}
-      <div class="text-muted">All checks ok{% if card.counts.ok %} ({{ card.counts.ok }}){% endif %} — <a href="{% url 'monitor_app:system_status' %}">System page</a>.</div>
-      {% endif %}
-    {% else %}
-      {% include 'monitor_app/_snapper_component_body.html' %}
-    {% endif %}
-
-    <details class="mt-1">
-      <summary class="text-muted small">component document (audit)</summary>
-      <pre class="p-2 border rounded" style="max-height:22em; overflow:auto; font-size:.9em;">{{ card.payload_json }}</pre>
-    </details>
-  </div>
-  {% endfor %}
-</div>
-{% if references %}
-<div class="mt-2">
-  <strong>Exact events around this instant:</strong>
-  {% for ref in references %}
-  {% if ref.page_url %}
-  <a class="me-3" href="{{ ref.page_url }}" title="{{ ref.transport.notes }}">{{ ref.page_label }}</a>
-  {% elif ref.transport.rest_url %}
-  <a class="me-3" href="{{ ref.transport.rest_url }}" target="_blank"
-     title="{{ ref.transport.notes }} Params: {{ ref.transport.rest_params }}">{{ ref.component }} ({{ ref.event_kind }})</a>
-  {% endif %}
-  {% endfor %}
-</div>
-{% endif %}
-{% endif %}
