+++ src/templates/base.html
@@ -471,7 +471,7 @@
                 <a href="{% url 'monitor_app:system_status' %}"
                    id="system-status-dropdown-link"
                    class="{% if system_status_overall == 'error' %}nav-system-error{% endif %}{% if active_nav.system %} nav-active{% endif %}">System Status</a>
-                <a href="{% url 'monitor_app:snapper_report' scope='epicprod' %}"
+                <a href="{% url 'snapper_ai:snapper_report' scope='epicprod' %}"
                    class="{% if active_nav.snapper %}nav-active{% endif %}">Snapper</a>
                 <a href="{% url 'canary:canary_page' %}"
                    class="{% if active_nav.canary %}nav-active{% endif %}">Canary</a>
